requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
redis==5.0.3
rq==1.15.1
python-multipart==0.0.9
pillow==10.2.0